backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
langchain==0.1.16
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-community==0.0.34
pinecone-client==3.2.2
groq==0.5.0
beautifulsoup4==4.12.3
requests==2.31.0
pdfplumber==0.11.0
pytest==8.1.1
